modify, delete 작업 완료 기타 디테일작업 미완

+++ src/pjh/ReViewDao.java
@@ -51,7 +51,7 @@ public class ReViewDao {
 			
 			vo = sqlSession.selectOne("review.rev_view", serial);
 			List<ReViewAttVo> attList = sqlSession.selectList("review.revAtt_view", serial);
-			
+			sqlSession.commit();
 			vo.setAttList(attList);
 			
 		}catch(Exception ex) {
@@ -108,24 +108,29 @@ public class ReViewDao {
 		
 		try {
 			r = sqlSession.update("review.rev_update", vo);
-             
+			sqlSession.commit();
+
 			if(r < 1) throw new Exception();
 			chkCnt = 0;
-			for(ReViewAttVo v : vo.getAttList()) {
-				v.setpSerial(vo.getSerial());
-				chkCnt += sqlSession.insert("review.revAtt_insert", v);
-			}
-			
-			if(chkCnt == vo.getAttList().size()) {
-				for(ReViewAttVo delVo : vo.getDelList()) {
-					sqlSession.delete("review.revAtt_delete2", delVo.getSysAtt());
-					File f = new File(ReViewFileUpload.saveDir + delVo.getSysAtt());
-					if(f.exists()) f.delete();
+			if(vo.getAttList() != null) {
+				for(ReViewAttVo v : vo.getAttList()) {
+					System.out.println("chkcnt " +chkCnt);
+					v.setpSerial(vo.getSerial());
+					chkCnt += sqlSession.insert("review.revAtt_insert", v);
 				}
-				sqlSession.close();
-			}else {
-				throw new Exception();
-			}
+				
+				if(chkCnt == vo.getAttList().size()) {
+					for(ReViewAttVo delVo : vo.getDelList()) {
+						sqlSession.delete("review.revAtt_delete2", delVo.getSysAtt());
+						sqlSession.commit();
+						File f = new File(ReViewFileUpload.saveDir + delVo.getSysAtt());
+						if(f.exists()) f.delete();
+					}
+					sqlSession.close();
+				}else {
+					throw new Exception();
+				}
+			} //첨부파일 유무로 수정
 		}catch(Exception ex) {
 			ex.printStackTrace();
 			System.out.println("r=" + r);
@@ -139,6 +144,38 @@ public class ReViewDao {
 			}
 		}
 		sqlSession.close();
+		return msg;
+	}
+	
+	public String delete(ReViewVo vo) {
+		String msg = "삭제가 완료되었습니다.";
+		List<ReViewAttVo> delList = null;
+		
+			try {
+				delList = sqlSession.selectList("review.rev_att_list",vo.getSerial());
+				System.out.println("delList size = " + delList.size());
+				
+				int r = sqlSession.delete("review.rev_delete", vo);
+				System.out.println("r = " + r);
+				if(r > 0) {
+					r = sqlSession.delete("review.revAtt_delete", vo.getSerial());
+					if(vo.getAttList() == null || r == vo.getAttList().size()) {
+						sqlSession.commit();
+						for(ReViewAttVo v : delList) {
+							File f = new File(ReViewFileUpload.saveDir + v.getSysAtt());
+							if(f.exists()) f.delete();
+						}
+					}else {
+						throw new Exception();
+					}
+				}else {
+					throw new Exception();
+				}
+			}catch(Exception ex) {
+				sqlSession.rollback();
+				ex.printStackTrace();
+				msg = ex.toString();
+			}
 		return msg;
 	}
